Add interaction map update during dialog pause
...and some other experiments to animate objects, too, that failed.

I suspect it may generally work but:

- the address I picked to store the interaction object was probably used
  for something
- registers were probably overwritten by the update routine

It may be problematic in non-interactions. In these at least, we know
the player has a choice of interacting or not, therefore it's perfectly
possible the objects would be animating anyway. Hence it should
logically be fine to keep one still while we talk to it. One exception
would be if the interaction turns into a scene which animates other
objects.

One method that might work is to put a flag on objects which should not
animate, similar to the do not interact render flag. We could add this
to objects which we needed to control during scenes. It may not be a
problem, though, since objects we need to control have their routine
overwritten to one that wouldn't move anyway.